The json serializer was accidentally fetching the db.ReferenceProperty objects with each object that it serialized.  This made what should have been a cheap datastore lookup + JSON serialization O(n).  Yucko!
The JSON serializer still can do either "serialize references" or just
serialize keys, but the serialize just keys case is fast now.  So, for
getting a single item from the store, serializing references is convenient
and cheap.  But for large fetches from the store, it is now much faster.

Thanks appstats for finding this!

Review URL: http://codereview.chromium.org/6362003

git-svn-id: http://web-page-replay.googlecode.com/svn/trunk@210 164481a5-15cb-f69f-4b93-856c5b7754c5