Server port change had a bug - it was now shaping the port to the backend server, not the port to the frontend server.
Add a flag to the replay.py script so that we can pass both a port to listen
on as well as a port to do shaping on.

Probably we should move the trafficshaper control out of the replay script
and into the runner, but that is a big deal.  And the trafficshaper self-
resets on start, so you can't use it both from runner and replay without
having it clobber itself (e.g. one would do shaping, the other, being off
would turn off shaping!)

TBR=tonyg

Review URL: http://codereview.chromium.org/6592030

git-svn-id: http://web-page-replay.googlecode.com/svn/trunk@335 164481a5-15cb-f69f-4b93-856c5b7754c5