Problems of multiple headers with the same name in an http response.
This CL is to make WPR handle multiple headers with the same name and fix send_error in
httpproxy.py.

The http response may contain multiple headers with the same name. For example,
there might be multiple "set-cookie" headers.
The correct format of these headers in the response is as bellow:

    set-cookie: GMAIL_IMP=EXPIRED; Expires=Wed, 11-Jul-2012 21:53:17 GMT;
Path=/mail; Secure
    set-cookie: GMAIL_STAT_3d81=EXPIRED; Expires=Wed, 11-Jul-2012 21:53:17 GMT;
Path=/mail; Secure

where each cookie got set by one "set-cookie" header.

However, it is observed from the client(browser) side that, multile "set-cookie"
headers becomes one line in the response from WPR, which looks like:
    set-cookie: GMAIL_IMP=EXPIRED; Expires=Wed, 11-Jul-2012 21:53:17 GMT;
Path=/mail; Secure, GMAIL_STAT_3d81=EXPIRED; Expires=Wed, 11-Jul-2012 21:53:17
GMT; Path=/mail; Secure

This bad formatting makes the client (browser) not be able to parse the
set-cookie headers properly.

The cause of this problem can be traced from httpclient.RealHttpFetch.__call__,
where it constructs the ArchivedHttpResponse using response.getheaders().

 archived_http_response = httparchive.ArchivedHttpResponse(
            response.version,
            response.status,
            response.reason,
            response.getheaders(),
            chunks,
            delays)

By looking into the value of response.getheaders, I found that the "set-cookie"
headers are formatted wrongly. I think this is a glitch in
httplib.HttpResponse.
(This issue is reported at http://bugs.python.org/issue1660009 and
http://stackoverflow.com/questions/1649401/how-to-handle-multiple-set-cookie-...)

The workaround is not to use response.getheaders() but use response.msg.headers
instead. I modified the code as such and I am able to log into gmail and other
website which requires logging in.

Another modification I made is in httpproxy.HttpArchiveHandler.send_error(self,
status). This function override the one in
BaseHTTPServer.BaseHTTPRequestHandler.
However, the signature of the function is not compatible with the one in its
parent class. This is problematic because other functions in the parent class
still make calls to this fucntion.
For example, BaseHTTPServer.BaseHTTPRequestHandler.parse_request makes call to
it via "self.send_error(400, "Bad request version (%r)" % version)".
So I believe we need to make the function signature look the same as it is in
its parent class.

BUG=None
TEST= httpclient_test.py
Review URL: https://chromiumcodereview.appspot.com/10807016

git-svn-id: http://web-page-replay.googlecode.com/svn/trunk@489 164481a5-15cb-f69f-4b93-856c5b7754c5