Add a deterministic getTimezoneOffset() method.
The real getTimezoneOffset() depends on the timezone of the machine. GMail
doesn't replay properly unless the timezone is the same as it was during the
recording. A deterministic implementation is necessary to replay on bots with
a timezone different than where the original recording was made.

This implementation fakes PST/PDT for the year 2010. That makes Closure's
TimeZoneDetector happy. For other years, the method is still deterministic, but
we don't bother to simulate DST there (as that would require hardcoding a
bunch of dates).
Review URL: https://codereview.chromium.org/11612029

git-svn-id: http://web-page-replay.googlecode.com/svn/trunk@506 164481a5-15cb-f69f-4b93-856c5b7754c5