Make WPR passthrough requests to other ports on the same machine.
This allows another web server on the same machine to serve requests. It is
useful for Telemetry which runs a static HTTP server for serving files on the
local filesystem. But chrome is running with --testing-fixed-http-port, so all
localhost requests get directed to WPR's port.

Anyway, we can now have Telemetry PageSets with both file:// and http:// URLs
where the http:// will be handled by WPR and file:// by the static server.

BUG=http://code.google.com/p/chromium/issues/detail?id=257209
R=slamm@google.com

Review URL: https://codereview.chromium.org/18713002

git-svn-id: http://web-page-replay.googlecode.com/svn/trunk@518 164481a5-15cb-f69f-4b93-856c5b7754c5