ISSUE-71: adjust DEFAULT_REQUEST_TIME to millisecond granularity.
There are pages which assume that the timestamp in deterministic date is exactly
the same during replay time as it was at the record time.  This change adjusts
default constant such that JS millisecond representation is exactly the same as
it was before when hardcoded into deterministic.js.

BUG=#71
R=nednguyen@google.com, nednguyen

Review URL: https://codereview.chromium.org/2545373002 .

Patch from Andrey Malets <malets@yandex-team.ru>.