Always include a subjectAltName and extendedKeyUsage extension in certs
The existing logic attempts to copy the subjectAltName from the existing
cert, if any is provided, but otherwise leaves it absent. It also omits
any other extensions.

This means that if no certificate is supplied to copy from, no
subjectAltName will be present on the leaf cert, which means it doesn't
work for clients that require the SAN always be present (because it's
the only unambiguous way to represent the hostname the cert is for).

This also updates the leaf to include the extendedKeyUsage of serverAuth
and clientAuth. Browsers currently have 'legacy' behaviours in which
they assume a certificate lacking an EKU is intended to be used as a
server certificate, but are working to always require it. This minimizes
any future compatibility risk.

In doing so, this removes the concept of the EXTENSION_WHITELIST added.
This is because copying extensions generically can be harmful, because
it depends on the context of the dummy cert. For example, copying
extensions like keyUsage or the TLS Feature Extension (aka OCSP
Must-Staple) would be actively harmful if WPR did not support the
ciphersuites used (in the case of keyUsage) or OCSP stapling (in the
case of TLS feature extension).